+++ api/data/books/book.py
@@ -1,6 +1,6 @@
 import datetime
 
-from flask import Response, jsonify, make_response
+from flask import Response
 from flask_restful import Resource
 
 from helpers.init import db
@@ -8,6 +8,9 @@ from helpers.jwt_auth import verify_jwt_token
 from helpers.request_parser import RequestParser
 from models import book
 from models.user import User
+from static.responses import create_response, TOKEN_INVALID_RESPONSE, INSUFFICIENT_PERMISSIONS_RESPONSE, \
+    OBJECT_MODIFIED_RESPONSE, OBJECT_DELETED_RESPONSE, OBJECT_CREATED_RESPONSE, BOOK_OBJECT_RESPONSE, \
+    BOOK_OBJECTS_LIST_RESPONSE
 
 
 class Book(Resource):
@@ -40,23 +43,18 @@ class Book(Resource):
     def get(self) -> Response:
         args: dict = self.get_parser.parse_args()
         book_id: int = args.get("id")
-        verification_output: Response | str = verify_jwt_token()
-        if type(verification_output) is str:
-            pass
-        else:
-            return make_response(verification_output, 401)
+        email: str | None = verify_jwt_token()
+        if not email:
+            return create_response(TOKEN_INVALID_RESPONSE)
 
         if book_id:
             book_object: book.Book = book.Book.query.filter_by(id=book_id).first()
-            return make_response(
-                jsonify(book_object.as_dict()),
-                200,
-            )
+            return create_response(BOOK_OBJECT_RESPONSE, book_object.as_dict()
+                                   )
         book_objects: list[book.Book] = book.Book.query.all()
-        return make_response(
-            jsonify(*[book_object.as_dict() for book_object in book_objects]),
-            200,
-        )
+        return create_response(
+            BOOK_OBJECTS_LIST_RESPONSE,
+            {"books": [book_object.as_dict() for book_object in book_objects]})
 
     def post(self) -> Response:
         args: dict = self.post_parser.parse_args()
@@ -68,20 +66,12 @@ class Book(Resource):
         genres: list[int] = args.get("genres")
         picture: str = args.get("picture")
         premiere_date: datetime.datetime = args.get("premiere_date")
-        verification_output: Response | str = verify_jwt_token()
-        if type(verification_output) is str:
-            email: str = verification_output
-        else:
-            return make_response(verification_output, 401)
+        email: str | None = verify_jwt_token()
+        if not email:
+            return create_response(TOKEN_INVALID_RESPONSE)
         user: User = User.query.filter_by(email=email).first()
         if not user.is_admin:
-            return make_response(
-                jsonify(
-                    message="insufficient_permissions",
-                    details="Insufficient permissions. Requires admin status.",
-                ),
-                404,
-            )
+            return create_response(INSUFFICIENT_PERMISSIONS_RESPONSE)
 
         # TODO A CO JEŚLI AUTOR O TAKIM ID NIE ISTNIEJE?
         book_object: book.Book = book.Book(
@@ -97,38 +87,24 @@ class Book(Resource):
         db.session.add(book_object)
         db.session.commit()
 
-        return make_response(
-            jsonify(message="book_created", details="Book created."),
-            200,
-        )
+        return create_response(OBJECT_CREATED_RESPONSE)
 
     def delete(self) -> Response:
         args: dict = self.delete_parser.parse_args()
         book_id: int = args.get("id")
-        verification_output: Response | str = verify_jwt_token()
-        if type(verification_output) is str:
-            email: str = verification_output
-        else:
-            return make_response(verification_output, 401)
+        email: str | None = verify_jwt_token()
+        if not email:
+            return create_response(TOKEN_INVALID_RESPONSE)
         user: User = User.query.filter_by(email=email).first()
         if not user.is_admin:
-            return make_response(
-                jsonify(
-                    message="insufficient_permissions",
-                    details="Insufficient permissions. Requires admin status.",
-                ),
-                404,
-            )
+            return create_response(INSUFFICIENT_PERMISSIONS_RESPONSE)
 
         opinion_object: book.Book = book.Book.query.filter_by(id=book_id).first()
 
         db.session.delete(opinion_object)
         db.session.commit()
 
-        return make_response(
-            jsonify(message="book_deleted", details="Book deleted."),
-            200,
-        )
+        return create_response(OBJECT_DELETED_RESPONSE)
 
     def patch(self) -> Response:
         args: dict = self.delete_parser.parse_args()
@@ -140,20 +116,12 @@ class Book(Resource):
         genres: list[int] = args.get("genres")
         picture: str = args.get("picture")
         premiere_date: datetime.datetime = args.get("premiere_date")
-        verification_output: Response | str = verify_jwt_token()
-        if type(verification_output) is str:
-            email: str = verification_output
-        else:
-            return make_response(verification_output, 401)
+        email: str | None = verify_jwt_token()
+        if not email:
+            return create_response(TOKEN_INVALID_RESPONSE)
         user: User = User.query.filter_by(email=email).first()
         if not user.is_admin:
-            return make_response(
-                jsonify(
-                    message="insufficient_permissions",
-                    details="Insufficient permissions. Requires admin status.",
-                ),
-                404,
-            )
+            return create_response(INSUFFICIENT_PERMISSIONS_RESPONSE)
 
         modified_book: book.Book = book.Book.query.filter_by(id=book_id).first()
 
@@ -174,7 +142,4 @@ class Book(Resource):
                 modified_book.premiere_date = premiere_date
             db.session.commit()
 
-        return make_response(
-            jsonify(message="book_modified", details="Book modified."),
-            200,
-        )
+        return create_response(OBJECT_MODIFIED_RESPONSE)
